+++ buy-me-coffee.php
@@ -4,7 +4,7 @@
 Plugin Name: Buy Me a Coffee button & widgets - Fundraise with Stripe and PayPal
 Plugin URI: https://wpminers.com/buymecoffee/
 Description: Easy way to collect donations like "buy me a coffee" directly your own Stripe and PayPal for free
-Version: 1.2.0
+Version: 1.2.1
 Author: wpminers
 Author URI: http://www.wpminers.com/
 License: GPLv2 or later
@@ -34,12 +34,12 @@ if (!defined('ABSPATH')) {
 }
 if (!defined('BUYMECOFFEE_VERSION')) {
     define('BUYMECOFFEE_VERSION_LITE', true);
-    define('BUYMECOFFEE_VERSION', '1.2.0');
+    define('BUYMECOFFEE_VERSION', '1.2.1');
     define('BUYMECOFFEE_MAIN_FILE', __FILE__);
     define('BUYMECOFFEE_URL', plugin_dir_url(__FILE__));
     define('BUYMECOFFEE_DIR', plugin_dir_path(__FILE__));
     define('BUYMECOFFEE_UPLOAD_DIR', '/buy-me-coffee');
-    define('BUYMECOFFEE_DEVELOPMENT', 'yes');
+    define('BUYMECOFFEE_PRODUCTION', 'yes');
     define('BUYMECOFFEE_DB_VERSION', '1.7');
 
     class BuyMeCoffee